requirements.txt
apache-airflow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
openpyxl
pandas
playwright
rapidfuzz